requirements.txt
asgiref==3.7.2
Django==4.2.7
<<<<<<< Updated upstream
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg==3.1.8
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
=======
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-nine==0.2.7
django-rest-registration==0.8.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
google-api-core==2.14.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
idna==3.4
inflection==0.5.1
kombu==5.3.4
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg==3.1.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
>>>>>>> Stashed changes
